+++ core/src/main/java/hudson/model/Run.java
@@ -121,6 +121,7 @@ import jenkins.model.Jenkins;
 import jenkins.model.JenkinsLocationConfiguration;
 import jenkins.model.RunAction2;
 import jenkins.model.StandardArtifactManager;
+import jenkins.model.Tab;
 import jenkins.model.TransientActionFactory;
 import jenkins.model.details.CauseDetail;
 import jenkins.model.details.Detail;
@@ -129,7 +130,6 @@ import jenkins.model.details.DurationDetail;
 import jenkins.model.details.TimestampDetail;
 import jenkins.model.lazy.BuildReference;
 import jenkins.model.lazy.LazyBuildMixIn;
-import jenkins.run.Tab;
 import jenkins.security.MasterToSlaveCallable;
 import jenkins.security.stapler.StaplerNotDispatchable;
 import jenkins.util.SystemProperties;
