+++ core/src/main/java/jenkins/run/ConsoleTabFactory.java
@@ -7,6 +7,7 @@ import hudson.model.Run;
 import java.util.Collection;
 import java.util.Collections;
 import jenkins.console.DefaultConsoleUrlProvider;
+import jenkins.model.Tab;
 import jenkins.model.TransientActionFactory;
 import jenkins.model.experimentalflags.NewBuildPageUserExperimentalFlag;
